chore: update naming from PR feedback
